Pull Job class into RequestQueue as Request
Since it is very tightly coupled to the queue anyway and it is not much
code.
